Refix

+++ Laboratory-3/pi_taylor_parallel.cc
@@ -8,7 +8,7 @@
 #include <vector>
 #include <chrono>
 
-using my_float = float;
+using my_float = long double;
 
 
 void pi_taylor_chunk(std::vector<my_float> &output, size_t thread_id, size_t start_step, size_t stop_step) {
